+++ lib/Controller/DuplicateApiController.php
@@ -50,7 +50,6 @@ class DuplicateApiController extends AbstractAPIController
     public function list(int $page = 1, int $limit = 30, string $type = 'unacknowledged'): DataResponse
     {
         try {
-            $offset = ($page - 1) * $limit; // Calculate the offset based on the current page and limit
             $duplicates = $this->fileDuplicateService->findAll($type, $this->getUserId(), $page, $limit, true);
             $totalItems = $this->fileDuplicateService->getTotalCount($type); 
             $totalPages = ceil($totalItems / $limit);