+++ lib/Controller/PageController.php
@@ -1,21 +1,41 @@
 <?php
 namespace OCA\DuplicateFinder\Controller;
 
-use OCP\IRequest;
-use OCP\AppFramework\Http\TemplateResponse;
+use OCA\Viewer\Event\LoadViewer;
 use OCP\AppFramework\Controller;
-use OCA\DuplicateFinder\AppInfo\Application;
+use OCP\AppFramework\Http\TemplateResponse;
+use OCP\EventDispatcher\IEventDispatcher;
+use OCP\IRequest;
 
 class PageController extends Controller
 {
+    protected $appName;
+
+    /** @var IEventDispatcher */
+    private $eventDispatcher;
+
+    public function __construct(
+        $appName,
+        IRequest $request,
+        IEventDispatcher $eventDispatcher
+    ) {
+        parent::__construct($appName, $request);
+
+        $this->appName = $appName;
+        $this->eventDispatcher = $eventDispatcher;
+    }
 
     /**
-     *
      * @NoAdminRequired
      * @NoCSRFRequired
+     * Render default index template
+     *
+     * @return TemplateResponse
      */
     public function index(): TemplateResponse
     {
-        return new TemplateResponse(Application::ID, 'App');
+        $this->eventDispatcher->dispatch(LoadViewer::class, new LoadViewer());
+        $response = new TemplateResponse($this->appName, 'App');
+        return $response;
     }
 }